move stencil test to stencil_test folder for proper test collection
